Added geolocation comparison and functionality
