+++ frontend/src/components/Geolocation.jsx
@@ -1,19 +1,71 @@
 import React, { useState, useEffect } from 'react';
 
-const GeolocationComponent = () => {
+const GeolocationComponent = ({ address }) => {
   const [locationJSON, setLocationJSON] = useState('');
+  const [distance, setDistance] = useState(null);
   const [error, setError] = useState(null);
+
+  // Function to geocode an address using Nominatim API
+  const geocodeAddress = async (address) => {
+    const url = `https://nominatim.openstreetmap.org/search?format=json&q=${encodeURIComponent(address)}`;
+    try {
+      const response = await fetch(url);
+      const data = await response.json();
+      if (data && data.length > 0) {
+        const { lat, lon } = data[0];
+        return { latitude: parseFloat(lat), longitude: parseFloat(lon) };
+      } else {
+        throw new Error("No results found for the provided address.");
+      }
+    } catch (error) {
+      console.error("Error geocoding address:", error);
+      throw error;
+    }
+  };
+
+  // Haversine formula to calculate distance between two coordinates (in km)
+  const haversineDistance = (lat1, lon1, lat2, lon2) => {
+    const toRad = (value) => (value * Math.PI) / 180;
+    const R = 6371; // Earth's radius in kilometers
+    const dLat = toRad(lat2 - lat1);
+    const dLon = toRad(lon2 - lon1);
+    const a =
+      Math.sin(dLat / 2) ** 2 +
+      Math.cos(toRad(lat1)) *
+        Math.cos(toRad(lat2)) *
+        Math.sin(dLon / 2) ** 2;
+    const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+    return R * c;
+  };
 
   useEffect(() => {
     if (navigator.geolocation) {
       navigator.geolocation.getCurrentPosition(
-        (position) => {
+        async (position) => {
           const { latitude, longitude } = position.coords;
-          // Create a JSON string from the latitude and longitude
+          // Create a JSON string from the user's latitude and longitude
           const locationData = { latitude, longitude };
           const jsonString = JSON.stringify(locationData);
           setLocationJSON(jsonString);
-          console.log("Location JSON:", jsonString);
+          console.log("User Location JSON:", jsonString);
+
+          try {
+            // Use the unique address prop from the parent
+            const addressCoords = await geocodeAddress(address);
+            console.log("Address Coordinates:", addressCoords);
+            // Calculate the distance between the user and the address
+            const calculatedDistance = haversineDistance(
+              latitude,
+              longitude,
+              addressCoords.latitude,
+              addressCoords.longitude
+            );
+            setDistance(calculatedDistance);
+            console.log(`Distance to ${address}: ${calculatedDistance.toFixed(2)} km`);
+          } catch (err) {
+            console.error("Error processing address:", err);
+            setError("Error geocoding address or calculating distance.");
+          }
         },
         (error) => {
           setError(error.message);
@@ -22,18 +74,25 @@ const GeolocationComponent = () => {
     } else {
       setError("Geolocation is not supported by this browser.");
     }
-  }, []);
+  }, [address]); // Effect runs again if the address prop changes
 
   return (
     <div>
       {error ? (
         <p>Error: {error}</p>
       ) : (
-        <p>
-          {locationJSON
-            ? `Location JSON: ${locationJSON}`
-            : "Fetching location..."}
-        </p>
+        <>
+          <p>
+            {/* {locationJSON
+              ? `User Location JSON: ${locationJSON}`
+              : "Fetching user location..."} */}
+          </p>
+          <p>
+            {distance !== null
+              ? `Distance to ${address}: ${distance.toFixed(2)} km`
+              : "Calculating distance..."}
+          </p>
+        </>
       )}
     </div>
   );
